pyproject.toml
[build-system]
requires = ["setuptools>=75.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sovereign-ink"
version = "0.1.0"
description = "Autonomous historical-political novel generation system"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.42.0",
    "pydantic>=2.10.0",
    "rich>=13.9.0",
    "questionary>=2.1.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    "tiktoken>=0.8.0",
    "click>=8.1.0",
    "schedule>=1.2.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.3.0"]

[project.scripts]
sovereign-ink = "sovereign_ink.main:main"

[tool.setuptools.packages.find]
include = ["sovereign_ink*"]
